Renamed files to match new type names
