+++ src/jolt_body_3d.cpp
@@ -1,10 +1,10 @@
-#include "jolt_physics_body_3d.hpp"
+#include "jolt_body_3d.hpp"
 
-#include "body_access.hpp"
 #include "conversion.hpp"
 #include "error_macros.hpp"
+#include "jolt_body_access_3d.hpp"
 #include "jolt_physics_direct_body_state_3d.hpp"
-#include "jolt_physics_space_3d.hpp"
+#include "jolt_space_3d.hpp"
 #include "utility_functions.hpp"
 
 JoltBody3D::~JoltBody3D() {
