+++ src/jolt_body_3d.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "jolt_physics_collision_object_3d.hpp"
+#include "jolt_collision_object_3d.hpp"
 #include "jolt_physics_direct_body_state_3d.hpp"
 
 class JoltBody3D final : public JoltCollisionObject3D {
